fix: rename folder + update comments
