+++ src/app/pages/prescription/stp1-patient-selection/stp1-patient-selection.component.css
