+++ src/app/pages/prescription/stp2-patient-details/stp2-patient-details.component.html
